--- PowerEditor/src/MISC/sha1/cal_sha1.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void calc_sha1(unsigned char hash[20], const void *input, size_t len);
